refactor: extract result aggregation function to separate hook
