+++ apps/frontend-pwa/src/pages/microlearning/[id]/evaluation.tsx
@@ -1,19 +1,17 @@
 import { useMutation, useQuery } from '@apollo/client'
 import Layout from '@components/Layout'
+import useStackEvaluationAggregation from '@components/hooks/useStackEvaluationAggregation'
 import {
-  ElementType,
   GetMicrolearningDocument,
   GetParticipationDocument,
   MarkMicrolearningCompletedDocument,
   SelfDocument,
 } from '@klicker-uzh/graphql/dist/ops'
 import Loader from '@klicker-uzh/shared-components/src/Loader'
-import { StudentResponseType } from '@klicker-uzh/shared-components/src/StudentElement'
 import { Button, H3, UserNotification } from '@uzh-bf/design-system'
 import { GetStaticPropsContext } from 'next'
 import { useTranslations } from 'next-intl'
 import { useRouter } from 'next/router'
-import { useMemo } from 'react'
 
 function MicrolearningEvaluation() {
   const t = useTranslations()
@@ -35,82 +33,7 @@ function MicrolearningEvaluation() {
   )
 
   const microlearning = data?.microlearning
-
-  // TODO: extract to separte hook
-  const aggregatedResults = useMemo(() => {
-    if (!microlearning) {
-      return { evaluation: undefined, totalPointsAwarded: 0 }
-    }
-
-    return microlearning.stacks?.reduce<{
-      evaluation: Record<
-        number,
-        { pointsAwarded: number; score: number; maxPoints: number }
-      >
-      totalPointsAwarded: number
-    }>(
-      (acc, stack) => {
-        const rawStackStorage = localStorage.getItem(
-          `qi-${microlearning.id}-${stack.id}`
-        )
-        const stackStorage: StudentResponseType = rawStackStorage
-          ? JSON.parse(rawStackStorage)
-          : null
-
-        // if no results for a stack are found, return 0 points
-        if (stackStorage === null) {
-          return {
-            evaluation: {
-              ...acc.evaluation,
-              [stack.id]: { pointsAwarded: 0, score: 0, maxPoints: 0 },
-            },
-            totalPointsAwarded: acc.totalPointsAwarded,
-          }
-        }
-
-        const stackResults = Object.values(stackStorage).reduce<{
-          stackPointsAwarded: number
-          stackScore: number
-          stackMaxPoints: number
-        }>(
-          (acc, entry) => {
-            // flashcards and content elements do not contribute to the score at the moment
-            if (
-              entry.type === ElementType.Flashcard ||
-              entry.type === ElementType.Content
-            ) {
-              return acc
-            }
-            // aggregate the awarded points over the questions
-            return {
-              stackPointsAwarded:
-                acc.stackPointsAwarded + (entry.evaluation?.pointsAwarded ?? 0),
-              stackScore: acc.stackScore + (entry.evaluation?.score ?? 0),
-              stackMaxPoints:
-                acc.stackMaxPoints +
-                (entry.evaluation?.pointsMultiplier ?? 1) * 10,
-            }
-          },
-          { stackPointsAwarded: 0, stackScore: 0, stackMaxPoints: 0 }
-        )
-
-        // combine date over stacks
-        return {
-          evaluation: {
-            ...acc.evaluation,
-            [stack.id]: {
-              pointsAwarded: stackResults.stackPointsAwarded,
-              score: stackResults.stackScore,
-              maxPoints: stackResults.stackMaxPoints,
-            },
-          },
-          totalPointsAwarded:
-            acc.totalPointsAwarded + stackResults.stackPointsAwarded,
-        }
-      },
-      { evaluation: {}, totalPointsAwarded: 0 }
-    )
-  }, [microlearning])
+  const aggregatedResults = useStackEvaluationAggregation({ microlearning })
 
   if (loading || !microlearning) {
     return (
